+++ examples/parallel_for/seismic/main.cpp
@@ -1,5 +1,5 @@
 /*
-    Copyright (c) 2005-2024 Intel Corporation
+    Copyright (c) 2005-2025 Intel Corporation
 
     Licensed under the Apache License, Version 2.0 (the "License");
     you may not use this file except in compliance with the License.
@@ -22,6 +22,7 @@
 #include "oneapi/tbb/global_control.h"
 
 #include "common/utility/utility.hpp"
+#include "common/utility/measurements.hpp"
 #include "common/utility/get_default_num_threads.hpp"
 
 #include "seismic_video.hpp"
